+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Annotations: NEW, IMProved, FIX, DEPrecated, REMoved, SECurity
 ## [0.0.0] - Unreleased
 
 - [NEW] - Initial release. 
-- [NEW] - Supports reading all [RIFF1991](https://wavref.til.cafe/spec/riff1991/) metadata chunks: RIFF, LIST, CSET, fmt, data, adtl, info, cue, fact, plst. 
+- [NEW] - Supports reading all [RIFF1991](https://wavref.til.cafe/spec/riff1991/) metadata chunks: RIFF, LIST-INFO, CSET, JUNK, fmt, data, LIST-adtl, info, cue, fact, plst, LIST-wavl, slnt. 
 - [NEW] - bext chunk support. [Broadcast Wave Format](https://en.wikipedia.org/wiki/Broadcast_Wave_Format)
 - [NEW] - MD5 of audio data chunk support. As specified by [BWFMetaEdit's Audio Data Checksums](https://mediaarea.net/BWFMetaEdit/md5)